+++ test/libsolidity/syntaxTests/nameAndTypeResolution/264_mapping_in_memory_array.sol
@@ -1,7 +1,7 @@
 contract C {
     function f(uint size) public {
-        mapping(uint => uint) x = new mapping(uint => uint)[](4);
+        mapping(uint => uint) storage x = new mapping(uint => uint)[](4);
     }
 }
 // ----
-// TypeError: (86-109): Type cannot live outside storage.
+// TypeError: (94-117): Type cannot live outside storage.