+++ Changelog.md
@@ -69,6 +69,7 @@ Bugfixes:
  * Tests: Fix chain parameters to make ipc tests work with newer versions of cpp-ethereum.
  * Code Generator: Fix allocation of byte arrays (zeroed out too much memory).
  * Fix NatSpec json output for `@notice` and `@dev` tags on contract definitions.
+ * References Resolver: Enforce ``storage`` as data location for mappings.
  * Type Checker: Consider fixed size arrays when checking for recursive structs.
  * Type System: Allow arbitrary exponents for literals with a mantissa of zero.